requirements.txt
mutagen==1.47.0
pillow==10.3.0
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
python-mpd2==3.1.1
PyYAML==6.0.1
shiboken6==6.7.0
